updated client unit testing
 AND created another option for asserting URL
