+++ tests/integration/PostgrestClientTest.php
@@ -43,6 +43,42 @@ final class PostgrestClientTest extends TestCase
         $this->assertArrayHasKey('name', $result->data[0]);
         $this->assertArrayHasKey('created_at', $result->data[0]);
     }
+    public function testCanBeCreatedFromValidUrl(): void
+    {
+        $this->assertInstanceOf(
+            PostgrestClient::class,
+            new PostgrestClient('http://localhost:3000', [])
+        );
+    }
+
+    public function testCanSelectTable(): void
+    {
+        $result = $this->client->from('users');//->select()->url->__toString();
+        print_r($result);
+        ob_flush();
+        //assertSame($result, 'gpdefvsxamnscceccczu');
+    }
+
+    public function testCanSelectColumns(): void
+    {
+        
+
+        $this->assertSame($client->from('users')->select('id, name')->url->__toString(), 'http://localhost:3000/users?select=id%2Cname');
+    }
+
+    public function testCanRPC(): void
+    {
+        $client = new PostgrestClient('http://localhost:3000', []);
+
+        $this->assertSame($client->rpc('add_one')->select()->url->__toString(), 'http://localhost:3000/rpc/add_one');
+    }
+
+    public function testCanExecute(): void
+    {
+        $client = new PostgrestClient('http://localhost:3000', []);
+
+        $this->assertSame($client->from('users')->select()->execute()->status, 200);
+    }
 
     public static function providerArray(): array
     {
